thread: Improve wait logic, implement missing functionality other minor fixes

+++ src/system/libroot2/Thread.cpp
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2018-2020, Dario Casalinuovo. All rights reserved.
+ *  Copyright 2018-2026, Dario Casalinuovo. All rights reserved.
  *  Distributed under the terms of the LGPL License.
  */
 
@@ -10,8 +10,11 @@
 #include <string>
 #include <map>
 
+#include <dirent.h>
+#include <poll.h>
 #include <pthread.h>
 #include <stdlib.h>
+#include <string.h>
 #include <sys/syscall.h>
 #include <sys/types.h>
 #include <sys/wait.h>
@@ -24,10 +27,6 @@
 #include "Thread.h"
 
 #include "../kernel/nexus/nexus/nexus.h"
-
-
-// TODO remove
-extern int gLoadImageFD;
 
 
 namespace BKernelPrivate {
@@ -66,14 +65,19 @@ Thread::Thread()
 {
 	fThread = find_thread(NULL);
 	fThreadBlockSem = create_sem(0, "block_thread_sem");
+	if (fThreadBlockSem < 0) {
+		TRACE("Thread: Failed to create block semaphore: %ld\n", fThreadBlockSem);
+	}
 	gCurrentThread = this;
 }
 
 
 Thread::~Thread()
 {
-	//assert(gettid() == this->fThreadID);
-	delete_sem(fThreadBlockSem);
+	if (fThreadBlockSem >= 0) {
+		delete_sem(fThreadBlockSem);
+		fThreadBlockSem = -1;
+	}
 	// TODO: pthread_detach
 }
 
@@ -82,7 +86,9 @@ void
 Thread::ReinitChildAtFork()
 {
 	TRACE("Process %d reinit thread after fork\n", getpid());
+
 	gThreadsMap.clear();
+	gCurrentThread = NULL;
 
 	_Init();
 }
@@ -93,6 +99,12 @@ Thread::CreateThread()
 {
 	thread_id id = find_thread(NULL);
 	Thread* thread = new Thread();
+
+	if (thread->fThreadBlockSem < 0) {
+		TRACE("CreateThread: block sem creation failed\n");
+		delete thread;
+		return B_NO_MEMORY;
+	}
 
 	gThreadsMap.insert(std::make_pair(id, thread));
 	return id;
@@ -116,31 +128,49 @@ void*
 Thread::thread_run(void* data)
 {
 	CALLED();
+
+	data_wrap* threadData = (data_wrap*)data;
+
 	Thread::Lock();
 	int nexus = BKernelPrivate::Team::GetNexusDescriptor();
-	if (ioctl(nexus, NEXUS_THREAD_SPAWN, "thread name") < 0) {
-		status_t exitStatus = B_ERROR;
-		pthread_exit(&exitStatus);
+	if (nexus_io(nexus, NEXUS_THREAD_SPAWN, "thread name") < 0) {
+		TRACE("thread_run: NEXUS_THREAD_SPAWN failed\n");
+		threadData->tid = -1;
+		Thread::Unlock();
+		Thread::Unblock(threadData->father, B_ERROR);
 		return NULL;
 	}
-	data_wrap* threadData = (data_wrap*)data;
-	threadData->tid = CreateThread();
+
+	thread_id tid = CreateThread();
+	if (tid < 0) {
+		TRACE("thread_run: CreateThread failed\n");
+		threadData->tid = -1;
+		Thread::Unlock();
+		Thread::Unblock(threadData->father, B_ERROR);
+		return NULL;
+	}
+
+	threadData->tid = tid;
 	Thread::Unlock();
 
 	Thread::Unblock(threadData->father, B_OK);
-	gCurrentThread->Block(B_TIMEOUT, B_INFINITE_TIMEOUT);
 
-	threadData->func(threadData->data);
+	status_t blockResult = gCurrentThread->Block(B_TIMEOUT, B_INFINITE_TIMEOUT);
+	if (blockResult != B_OK) {
+		TRACE("thread_run: Block failed with %ld\n", blockResult);
+	}
+
+	status_t threadResult = threadData->func(threadData->data);
 
 	Thread::Lock();
 	gThreadsMap.erase(threadData->tid);
-	printf("%d is exiting\n", gettid());
+	TRACE("%d is exiting\n", gettid());
 	delete gCurrentThread;
 	gCurrentThread = NULL;
 	delete threadData;
 	Thread::Unlock();
 
-	exit_thread(B_OK);
+	exit_thread(threadResult);
 	return NULL;
 }
 
@@ -152,14 +182,56 @@ Thread::Resume(thread_id id)
 }
 
 
+#if 0
+
 status_t
 Thread::Block(uint32 flags, bigtime_t timeout)
 {
-	if (fThreadBlockSem == -1)
-		debugger("ThreadBlockSem uninitialized");
+	CALLED();
 
-	acquire_sem_etc(fThreadBlockSem, 1, flags, timeout);
-	return B_OK;
+	struct nexus_thread_exchange exchange;
+	memset(&exchange, 0, sizeof(exchange));
+	exchange.op = NEXUS_THREAD_BLOCK;
+	exchange.flags = flags;
+	exchange.timeout = timeout;
+
+	int sNexus = BKernelPrivate::Team::GetNexusDescriptor();
+	status_t ret = nexus_io(sNexus, NEXUS_THREAD_OP, &exchange);
+	return ret;
+}
+
+
+status_t
+Thread::Unblock(thread_id thread, status_t status)
+{
+	CALLED();
+
+	if (thread < 0)
+		return B_BAD_THREAD_ID;
+
+	struct nexus_thread_exchange exchange;
+	memset(&exchange, 0, sizeof(exchange));
+	exchange.op = NEXUS_THREAD_UNBLOCK;
+	exchange.receiver = thread;
+	exchange.status = status;
+
+	int sNexus = BKernelPrivate::Team::GetNexusDescriptor();
+	return nexus_io(sNexus, NEXUS_THREAD_OP, &exchange);
+}
+
+
+#else
+
+status_t
+Thread::Block(uint32 flags, bigtime_t timeout)
+{
+	if (fThreadBlockSem < 0) {
+		TRACE("Block: ThreadBlockSem uninitialized or invalid\n");
+		debugger("ThreadBlockSem uninitialized");
+		return B_ERROR;
+	}
+
+	return acquire_sem_etc(fThreadBlockSem, 1, flags, timeout);
 }
 
 
@@ -172,27 +244,36 @@ Thread::Unblock(thread_id thread, status_t status)
 		Thread::Unlock();
 		return B_BAD_THREAD_ID;
 	}
+
 	Thread* ret = elem->second;
 	sem_id blockSem = -1;
 	if (ret != NULL)
 		blockSem = ret->fThreadBlockSem;
 	Thread::Unlock();
 
-	if (blockSem != -1) {
-		release_sem(blockSem);
-		return B_OK;
+	if (blockSem >= 0) {
+		status_t result = release_sem(blockSem);
+		if (result != B_OK) {
+			TRACE("Unblock: release_sem failed with %ld\n", result);
+		}
+		return result;
 	}
 
 	return B_ERROR;
 }
+
+#endif
 
 
 void
 Thread::_Init()
 {
 	pthread_mutex_init(&gLock, NULL);
-	gLoadImageFD = -1;
-	CreateThread();
+
+	thread_id id = CreateThread();
+	if (id < 0) {
+		TRACE("Thread::_Init: CreateThread failed\n");
+	}
 }
 
 
@@ -207,21 +288,36 @@ spawn_thread(thread_func func, const char* name, int32 priority, void* data)
 {
 	CALLED();
 
+	if (func == NULL) {
+		return B_BAD_VALUE;
+	}
+
 	BKernelPrivate::data_wrap* dataWrap
 		= new BKernelPrivate::data_wrap();
-	//dataWrap->name
+	if (dataWrap == NULL)
+		return B_NO_MEMORY;
+
 	dataWrap->data = data;
 	dataWrap->func = func;
 	dataWrap->father = find_thread(NULL);
+	dataWrap->tid = -1;
 
 	pthread_t pThread;
 	int32 ret = pthread_create(&pThread, NULL,
 		BKernelPrivate::Thread::thread_run, dataWrap);
-	if (ret < 0)
+	if (ret != 0) {
+		delete dataWrap;
 		return -errno;
+	}
 
-	BKernelPrivate::gCurrentThread->Block(B_TIMEOUT,
+	status_t blockResult = BKernelPrivate::gCurrentThread->Block(B_TIMEOUT,
 		B_INFINITE_TIMEOUT);
+
+	if (blockResult != B_OK)
+		TRACE("spawn_thread: Block failed with %ld\n", blockResult);
+
+	if (dataWrap->tid < 0)
+		return B_ERROR;
 
 	return dataWrap->tid;
 }
@@ -241,43 +337,41 @@ rename_thread(thread_id thread, const char* newName)
 	CALLED();
 
 	if (newName == NULL)
-		return B_ERROR;
+		return B_BAD_VALUE;
 
 	struct nexus_thread_exchange exchange;
+	memset(&exchange, 0, sizeof(exchange));
 	exchange.op = NEXUS_THREAD_SET_NAME;
 	exchange.buffer = newName;
 	exchange.receiver = thread;
 
 	int nexus = BKernelPrivate::Team::GetNexusDescriptor();
-	if (ioctl(nexus, NEXUS_THREAD_OP, &exchange) < 0)
-		return B_BAD_THREAD_ID;
-
-	return B_OK;
+	return nexus_io(nexus, NEXUS_THREAD_OP, &exchange);
 }
 
 
 void
 exit_thread(status_t status)
 {
-	int nexus = BKernelPrivate::Team::GetNexusDescriptor();
-	// TODO pass exit status here
-	if (ioctl(nexus, NEXUS_THREAD_EXIT, NULL) < 0)
-		status = B_ERROR;
+	struct nexus_thread_exchange exchange;
+	memset(&exchange, 0, sizeof(exchange));
+	exchange.return_code = status;
 
-	// TODO what about the status here?
-	pthread_exit((void*)&status);
+	int nexus = BKernelPrivate::Team::GetNexusDescriptor();
+	nexus_io(nexus, NEXUS_THREAD_EXIT, &exchange);
+	pthread_exit((void*)(intptr_t)status);
 }
 
 
 status_t
 on_exit_thread(void (*callback)(void *), void* data)
 {
-    int ret = pthread_key_create(&BKernelPrivate::sOnExitKey, callback);
-    if (ret != 0)
+	int ret = pthread_key_create(&BKernelPrivate::sOnExitKey, callback);
+	if (ret != 0)
 		return B_ERROR;
 
-    ret = pthread_setspecific(BKernelPrivate::sOnExitKey, data);
-    if (ret != 0)
+	ret = pthread_setspecific(BKernelPrivate::sOnExitKey, data);
+	if (ret != 0)
 		return B_ERROR;
 
 	return B_OK;
@@ -290,10 +384,11 @@ send_data(thread_id thread, int32 code,
 {
 	CALLED();
 
-	if (buffer == NULL || buffer == NULL || bufferSize < 0)
+	if (buffer == NULL || bufferSize == 0)
 		return B_BAD_VALUE;
 
 	struct nexus_thread_exchange exchange;
+	memset(&exchange, 0, sizeof(exchange));
 	exchange.op = NEXUS_THREAD_WRITE;
 	exchange.buffer = buffer;
 	exchange.size = bufferSize;
@@ -301,10 +396,7 @@ send_data(thread_id thread, int32 code,
 	exchange.receiver = thread;
 
 	int nexus = BKernelPrivate::Team::GetNexusDescriptor();
-	if (ioctl(nexus, NEXUS_THREAD_OP, &exchange) < 0)
-		return B_ERROR;
-
-	return B_OK;
+	return nexus_io(nexus, NEXUS_THREAD_OP, &exchange);
 }
 
 
@@ -313,22 +405,25 @@ receive_data(thread_id* sender, void* buffer, size_t bufferSize)
 {
 	CALLED();
 
-	if (sender == NULL || buffer == NULL || bufferSize < 0)
+	if (sender == NULL || buffer == NULL || bufferSize == 0)
 		return B_BAD_VALUE;
 
 	struct nexus_thread_exchange exchange;
+	memset(&exchange, 0, sizeof(exchange));
 	exchange.op = NEXUS_THREAD_READ;
 	exchange.buffer = buffer;
 	exchange.size = bufferSize;
 
 	// TODO B_INTERRUPTED
 	int nexus = BKernelPrivate::Team::GetNexusDescriptor();
-	if (ioctl(nexus, NEXUS_THREAD_OP, &exchange) < 0)
-		return B_ERROR;
+	status_t ret = nexus_io(nexus, NEXUS_THREAD_OP, &exchange);
+	if (ret < 0)
+		return ret;
 
-	*sender = exchange.sender;
+	if (sender)
+		*sender = exchange.sender;
 
-	return exchange.return_code;
+	return ret;
 }
 
 
@@ -339,14 +434,13 @@ has_data(thread_id thread)
 		return false;
 
 	struct nexus_thread_exchange exchange;
+	memset(&exchange, 0, sizeof(exchange));
 	exchange.op = NEXUS_THREAD_HAS_DATA;
 	exchange.receiver = thread;
 
 	int nexus = BKernelPrivate::Team::GetNexusDescriptor();
-	if (ioctl(nexus, NEXUS_THREAD_OP, &exchange) < 0)
-		return false;
 
-	return exchange.return_code == B_OK;
+	return nexus_io(nexus, NEXUS_THREAD_OP, &exchange) == B_OK;
 }
 
 
@@ -355,31 +449,114 @@ _get_thread_info(thread_id id, thread_info* info, size_t size)
 {
 	CALLED();
 
-	if (id < 0 || info == NULL || size != sizeof(thread_info))
+	if (id < 0)
+		return B_BAD_THREAD_ID;
+
+	if (info == NULL || size != sizeof(thread_info))
 		return B_BAD_VALUE;
 
-	// TODO: Proc
+	char path[64];
+	char buffer[512];
 
+	snprintf(path, sizeof(path), "/proc/%d/comm", id);
+	int fd = open(path, O_RDONLY);
+	if (fd < 0)
+		return B_BAD_THREAD_ID;
+
+	memset(info, 0, sizeof(thread_info));
 	info->thread = id;
-	memset(&info->name, 0, B_OS_NAME_LENGTH);
-	strncpy(info->name, "Unknown", B_OS_NAME_LENGTH);
-	//info->name[B_OS_NAME_LENGTH - 1] = '\0';
+
+	ssize_t len = read(fd, buffer, sizeof(buffer) - 1);
+	close(fd);
+
+	if (len > 0) {
+		buffer[len] = '\0';
+		char* newline = strchr(buffer, '\n');
+		if (newline != NULL)
+			*newline = '\0';
+		strncpy(info->name, buffer, B_OS_NAME_LENGTH - 1);
+	} else {
+		strncpy(info->name, "unknown", B_OS_NAME_LENGTH - 1);
+	}
+
 	info->state = B_THREAD_RUNNING;
-	info->priority = 5;
-	info->team = getpid();
+	info->priority = 10;
+	info->team = id;
+
+	snprintf(path, sizeof(path), "/proc/%d/status", id);
+	fd = open(path, O_RDONLY);
+	if (fd >= 0) {
+		len = read(fd, buffer, sizeof(buffer) - 1);
+		close(fd);
+
+		if (len > 0) {
+			buffer[len] = '\0';
+
+			char* tgid = strstr(buffer, "Tgid:");
+			if (tgid != NULL)
+				info->team = atoi(tgid + 5);
+
+			char* state = strstr(buffer, "State:");
+			if (state != NULL) {
+				state += 6;
+				while (*state == ' ' || *state == '\t')
+					state++;
+				switch (*state) {
+					case 'R':
+						info->state = B_THREAD_RUNNING;
+						break;
+					case 'S':
+					case 'D':
+						info->state = B_THREAD_WAITING;
+						break;
+					case 'T':
+						info->state = B_THREAD_SUSPENDED;
+						break;
+					default:
+						info->state = B_THREAD_RUNNING;
+						break;
+				}
+			}
+		}
+	}
+
 	return B_OK;
 }
 
 
 status_t
-_get_next_thread_info(team_id team, int32* cookie,
-	thread_info* info, size_t size)
+_get_next_thread_info(team_id team, int32* cookie, thread_info* info, size_t size)
 {
-	if (info == NULL || cookie == NULL || size != sizeof(thread_info))
+	if (cookie == NULL || info == NULL || size != sizeof(thread_info))
 		return B_BAD_VALUE;
 
-	// TODO: Proc
-	UNIMPLEMENTED();
+	if (team == 0)
+		team = getpid();
+
+	char path[64];
+	snprintf(path, sizeof(path), "/proc/%d/task", team);
+
+	DIR* dir = opendir(path);
+	if (dir == NULL)
+		return B_BAD_TEAM_ID;
+
+	struct dirent* entry;
+	int32 index = 0;
+
+	while ((entry = readdir(dir)) != NULL) {
+		if (entry->d_name[0] == '.')
+			continue;
+
+		if (index == *cookie) {
+			thread_id tid = atoi(entry->d_name);
+			closedir(dir);
+			(*cookie)++;
+			return _get_thread_info(tid, info, size);
+		}
+		index++;
+	}
+
+	closedir(dir);
 	return B_BAD_VALUE;
 }
 
@@ -409,7 +586,59 @@ set_thread_priority(thread_id id, int32 priority)
 	// 10, SCHED_OTHER 0
 	// 11-30, SCHED_OTHER -1,-20
 	// 31-120, SCHED_FIFO, 1,99
-	return B_BAD_THREAD_ID;
+	return B_OK;
+}
+
+
+status_t
+wait_for_thread_posix(thread_id id, status_t* returnCode)
+{
+	CALLED();
+
+	if (id <= 0)
+		return B_BAD_VALUE;
+
+	errno = 0;
+	int pfd = syscall(SYS_pidfd_open, id, 0);
+	if (pfd < 0)
+		return pfd;
+
+	struct pollfd p = { .fd = pfd, .events = POLLIN };
+	int pollfd = poll(&p, 1, -1);
+	if (pollfd < 0) {
+		int saved = errno;
+		close(pfd);
+		if (saved == EINTR)
+			return B_INTERRUPTED;
+		return B_ERROR;
+	}
+
+	siginfo_t si;
+	memset(&si, 0, sizeof(si));
+	if (waitid(P_PIDFD, pfd, &si, WEXITED | WNOWAIT) != 0) {
+		int saved = errno;
+		close(pfd);
+		if (saved == ECHILD)
+			return B_ENTRY_NOT_FOUND;
+		return B_ERROR;
+	}
+
+	status_t status = B_ERROR;
+	if (si.si_code == CLD_EXITED) {
+		status = (status_t)si.si_status;
+	} else if (si.si_code == CLD_KILLED || si.si_code == CLD_DUMPED) {
+		status = (status_t)(-si.si_status);
+	} else {
+		status = B_ERROR;
+	}
+
+	waitid(P_PIDFD, pfd, &si, WEXITED);
+	close(pfd);
+
+	if (returnCode)
+		*returnCode = status;
+
+	return B_OK;
 }
 
 
@@ -418,18 +647,38 @@ wait_for_thread(thread_id id, status_t* returnCode)
 {
 	CALLED();
 
-	// TODO resume thread if was spawned just now
+	if (id < 0)
+		return B_BAD_THREAD_ID;
+
+	if (id == getpid()) {
+		debugger("Team::WaitForTeam: cannot wait for own team");
+		return B_BAD_TEAM_ID;
+	}
+
 	struct nexus_thread_exchange exchange;
+	memset(&exchange, 0, sizeof(exchange));
 	exchange.op = NEXUS_THREAD_WAITFOR;
 	exchange.receiver = id;
 
 	int nexus = BKernelPrivate::Team::GetNexusDescriptor();
-	if (ioctl(nexus, NEXUS_THREAD_OP, &exchange) < 0)
-		return BKernelPrivate::Team::WaitForTeam(id, returnCode);
+	status_t ret = nexus_io(nexus, NEXUS_THREAD_OP, &exchange);
 
-	// TODO fix this
+	if (ret == B_BAD_THREAD_ID) {
+		if (kill(id, 0) < 0)
+			return -errno;
+
+		kill(id, SIGCONT);
+		return wait_for_thread_posix(id, returnCode);
+	}
+
+	if (ret == B_INTERRUPTED)
+		return B_INTERRUPTED;
+
+	if (ret != B_OK)
+		return ret;
+
 	if (returnCode != NULL)
-		*returnCode = 0;
+		*returnCode = exchange.return_code;
 
 	return B_OK;
 }
@@ -464,7 +713,11 @@ snooze_until(bigtime_t time, int timeBase)
 	if (timeBase != B_SYSTEM_TIMEBASE)
 		return B_ERROR;
 
-	return snooze(time - system_time());
+	bigtime_t now = system_time();
+	if (time <= now)
+		return B_OK;
+
+	return snooze(time - now);
 }
 
 
@@ -472,6 +725,12 @@ status_t
 _kern_block_thread(uint32 flags, bigtime_t timeout)
 {
 	CALLED();
+
+	if (BKernelPrivate::gCurrentThread == NULL) {
+		TRACE("_kern_block_thread: no current thread!\n");
+		return B_ERROR;
+	}
+
 	return BKernelPrivate::gCurrentThread->Block(flags, timeout);
 }
